chore: moved shadcn primitives to /components/primitives and composites to /components/composites
