+++ src/lib/components/composites/navigation-bar/ProjectNavigationBar.svelte
