+++ src/lib/components/composites/navigation-bar/SimpleNavigationBar.svelte
