+++ src/lib/components/primitives/avatar/avatar-fallback.svelte
