+++ src/lib/components/primitives/avatar/avatar-image.svelte
