+++ src/lib/components/primitives/avatar/avatar.svelte
