+++ src/lib/components/primitives/avatar/index.ts
