+++ src/lib/components/primitives/card/card-description.svelte
